+++ Frontend/Xipona.Frontend.Redux/Items/States/EditedItemAvailability.cs
@@ -1,3 +1,3 @@
 ﻿namespace ProjectHermes.Xipona.Frontend.Redux.Items.States;
 
-public record EditedItemAvailability(Guid StoreId, Guid DefaultSectionId, float PricePerQuantity);
+public record EditedItemAvailability(Guid StoreId, Guid DefaultSectionId, decimal PricePerQuantity);
